# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Lab Launcher API",
    "version": "0.7.0",
    "description": "Pre-1.0 JSON API for the bundled frontend, trusted plugins, and private-deployment automation. Mutating API routes require the session CSRF token in the X-CSRF-Token header, except /api/bootstrap and /api/auth/login. Automation may instead authenticate with an Admin-issued API token via Authorization: Bearer; token requests skip CSRF, act with the token's role, and are rejected on session/profile endpoints (/api/auth/*, /api/me*, /api/bootstrap*)."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "security": [
    { "sessionCookie": [] },
    { "bearerToken": [] }
  ],
  "paths": {
...
    "/me/preferences": {
      "get": {
        "tags": ["Profile"],
        "summary": "Read current user's preferences",
        "responses": { "200": { "description": "Preferences", "content": { "application/json": { "schema": { "type": "object", "properties": { "preferences": { "$ref": "#/components/schemas/Preferences" } } } } } }, "401": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/me/preferences/{key}": {
      "put": {
        "tags": ["Profile"],
        "summary": "Save one preference",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/PreferenceKey" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["value"], "properties": { "value": true } } } } },
        "responses": { "200": { "$ref": "#/components/responses/Ok" }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      },
      "delete": {
        "tags": ["Profile"],
        "summary": "Delete one preference",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/PreferenceKey" }],
        "responses": { "200": { "$ref": "#/components/responses/Ok" }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/users": {
      "get": { "tags": ["Users"], "summary": "List users", "responses": { "200": { "description": "Users", "content": { "application/json": { "schema": { "type": "object", "properties": { "users": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } } } }, "403": { "$ref": "#/components/responses/Error" } } },
      "post": { "tags": ["Users"], "summary": "Create a user", "security": [{ "sessionCookie": [], "csrfToken": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserWriteRequest" } } } }, "responses": { "201": { "$ref": "#/components/responses/UserEnvelope" }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/users/{id}": {
      "patch": { "tags": ["Users"], "summary": "Update a user", "description": "Changing password, resetting TOTP, or changing role revokes target-user sessions. At least one Admin account must remain.", "security": [{ "sessionCookie": [], "csrfToken": [] }], "parameters": [{ "$ref": "#/components/parameters/Id" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserWriteRequest" } } } }, "responses": { "200": { "$ref": "#/components/responses/RevocationOk" }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } },
      "delete": { "tags": ["Users"], "summary": "Delete a user", "description": "Deleting a user revokes that user's sessions. At least one Admin account must remain.", "security": [{ "sessionCookie": [], "csrfToken": [] }], "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "200": { "$ref": "#/components/responses/RevocationOk" }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "hll.sid" },
      "csrfToken": { "type": "apiKey", "in": "header", "name": "X-CSRF-Token" },
      "bearerToken": { "type": "http", "scheme": "bearer", "description": "Admin-issued API token (hll_ prefix). Skips CSRF; not accepted on session/profile endpoints." }
    },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
      "PreferenceKey": { "name": "key", "in": "path", "required": true, "schema": { "type": "string", "enum": ["favorites", "launchpad", "plugins"] } }
    },
    "responses": {
      "Ok": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } },
      "RevocationOk": { "description": "Success with session revocation count", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RevocationOk" } } } },
      "Error": { "description": "Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "UserEnvelope": { "description": "User envelope", "content": { "application/json": { "schema": { "type": "object", "required": ["user"], "properties": { "user": { "$ref": "#/components/schemas/User" } } } } } },
...
    },
    "schemas": {
      "Ok": { "type": "object", "required": ["ok"], "properties": { "ok": { "type": "boolean" } }, "additionalProperties": true },
      "RevocationOk": { "allOf": [{ "$ref": "#/components/schemas/Ok" }, { "type": "object", "properties": { "revokedSessions": { "type": "integer", "minimum": 0 } } }] },
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } }, "additionalProperties": true },
...
      "Role": { "type": "string", "enum": ["admin", "editor", "user"] },
      "User": { "type": "object", "properties": { "id": { "type": "integer" }, "username": { "type": "string" }, "role": { "$ref": "#/components/schemas/Role" }, "totpEnabled": { "type": ["boolean", "integer"] }, "recoveryCodesRemaining": { "type": "integer", "description": "Unused TOTP recovery codes. Present on /api/me when 2FA is enabled." }, "createdAt": { "type": "string" }, "updatedAt": { "type": "string" } }, "additionalProperties": true },
...
      "Preferences": { "type": "object", "properties": { "favorites": { "type": "array", "items": { "type": "string" } }, "launchpad": { "$ref": "#/components/schemas/LaunchpadPreferences" }, "plugins": { "$ref": "#/components/schemas/PluginPreferences" } }, "additionalProperties": true },
      "PluginPreferences": { "type": "object", "additionalProperties": { "type": "object", "additionalProperties": { "oneOf": [{ "type": "string" }, { "type": "number" }, { "type": "boolean" }] } } },
      "LaunchpadPreferences": { "type": "object", "properties": { "hiddenCategories": { "type": "array", "items": { "type": "string" } }, "viewMode": { "type": "string", "enum": ["cards", "compact", "list"] }, "hideMetadata": { "type": "boolean" }, "layoutOrder": { "type": "array", "items": { "type": "string" } } }, "additionalProperties": false },
...
      "UserWriteRequest": { "type": "object", "properties": { "username": { "type": "string", "minLength": 3 }, "password": { "type": "string", "minLength": 10 }, "role": { "$ref": "#/components/schemas/Role" }, "resetTotp": { "type": "boolean" } }, "additionalProperties": false },
...
    }
  }
}